core, datasource/etheruem, graph, runtime/wasm, server/http: elastic logging
