+++ core/src/subgraph/instance_manager.rs
@@ -125,7 +125,8 @@ impl SubgraphInstanceManager {
                                         general: elastic_config,
                                         index: String::from("subgraph-logs"),
                                         document_type: String::from("log"),
-                                        subgraph_id: manifest.id.clone(),
+                                        custom_id_key: String::from("subgraphId"),
+                                        custom_id_value: manifest.id.clone().deref().to_string(),
                                         flush_interval: Duration::from_secs(5),
                                     },
                                     term_logger.clone(),
