+++ datasource/ethereum/src/block_ingestor.rs
@@ -27,12 +27,33 @@ where
         ancestor_count: u64,
         logger: Logger,
         polling_interval: Duration,
+        elastic_config: Option<ElasticLoggingConfig>,
     ) -> Result<BlockIngestor<S, E>, Error> {
+        let term_logger = logger.new(o!("component" => "BlockIngestor"));
+        let logger = elastic_config
+            .clone()
+            .map(|elastic_config| {
+                split_logger(
+                    term_logger.clone(),
+                    elastic_logger(
+                        ElasticDrainConfig {
+                            general: elastic_config,
+                            index: String::from("block-ingestor-logs"),
+                            document_type: String::from("log"),
+                            custom_id_key: String::from("componentName"),
+                            custom_id_value: String::from("blockIngestor"),
+                            flush_interval: polling_interval,
+                        },
+                        term_logger.clone(),
+                    ),
+                )
+            })
+            .unwrap_or(term_logger);
         Ok(BlockIngestor {
             chain_store,
             eth_adapter,
             ancestor_count,
-            logger: logger.new(o!("component" => "BlockIngestor")),
+            logger,
             polling_interval,
         })
     }
