+++ store/test-store/src/lib.rs
@@ -1,7 +1,7 @@
 use crate::tokio::runtime::Runtime;
+use graph::log;
 #[allow(unused_imports)]
 use graph::prelude::{Store as _, *};
-use graph::util::log;
 use graph::web3::types::H256;
 use graph_store_postgres::{Store, StoreConfig};
 use lazy_static::lazy_static;
